mythic-index/MemoryQuill/story-content/chapter-location-analyzer/requirements.txt
# Imagery Pipeline Dependencies

# Anthropic API client
anthropic

# CLI framework
typer[all]

# Logging
loguru

# YAML parsing (included in stdlib, but explicit for clarity)
pyyaml
